feat: 用户维度限流调整 #1539

+++ src/backend/common/common-ratelimiter/src/main/kotlin/com/tencent/bkrepo/common/ratelimiter/algorithm/DistributedTokenBucketRateLimiter.kt
@@ -34,7 +34,7 @@ import org.springframework.data.redis.core.script.DefaultRedisScript
 
 class DistributedTokenBucketRateLimiter(
     private val key: String,
-    private val permitsPerSecond: Long,
+    private val permitsPerSecond: Double,
     private val capacity: Long,
     private val redisTemplate: RedisTemplate<String, String>,
     private val permits: Long = 1L,
